+++ src/main/java/dev/xeonkryptos/xeonrobotframeworkplugin/psi/ref/RobotKeywordFileResolver.java
@@ -1,16 +1,9 @@
 package dev.xeonkryptos.xeonrobotframeworkplugin.psi.ref;
 
-import com.intellij.openapi.module.Module;
-import com.intellij.openapi.module.ModuleUtilCore;
 import com.intellij.openapi.progress.ProcessCanceledException;
 import com.intellij.openapi.progress.ProgressManager;
-import com.intellij.openapi.projectRoots.Sdk;
-import com.intellij.openapi.roots.OrderRootType;
 import com.intellij.openapi.util.Key;
-import com.intellij.openapi.vfs.VfsUtilCore;
-import com.intellij.openapi.vfs.VirtualFile;
 import com.intellij.psi.PsiElement;
-import com.intellij.psi.PsiFile;
 import com.intellij.psi.util.CachedValue;
 import com.intellij.psi.util.CachedValueProvider.Result;
 import com.intellij.psi.util.CachedValuesManager;
@@ -23,7 +16,6 @@ import com.jetbrains.python.psi.PyFile;
 import com.jetbrains.python.psi.PyFunction;
 import com.jetbrains.python.psi.PyTargetExpression;
 import com.jetbrains.python.psi.types.TypeEvalContext;
-import com.jetbrains.python.sdk.PythonSdkUtil;
 import dev.xeonkryptos.xeonrobotframeworkplugin.MyLogger;
 import dev.xeonkryptos.xeonrobotframeworkplugin.ide.config.RobotOptionsProvider;
 import dev.xeonkryptos.xeonrobotframeworkplugin.psi.dto.KeywordDto;
@@ -53,7 +45,6 @@ class RobotKeywordFileResolver {
     private static final Key<CachedValue<Collection<DefinedVariable>>> VARIABLES_CACHE_KEY = new Key<>("ROBOT_PYTHON_FILE_VARIABLES_CACHE");
     private static final Key<CachedValue<Collection<DefinedVariable>>> FILE_VARIABLES_CACHE_KEY = new Key<>("ROBOT_PYTHON_FILE_VARIABLES_CACHE");
     private static final Key<CachedValue<Collection<DefinedVariable>>> CLASS_VARIABLES_CACHE_KEY = new Key<>("ROBOT_PYTHON_CLASS_VARIABLES_CACHE");
-    private static final Key<CachedValue<Boolean>> SYSTEM_PSI_FILE_KEY = new Key<>("ROBOT_PYTHON_SYSTEM_FILE_CACHE");
 
     private RobotKeywordFileResolver() {
         throw new UnsupportedOperationException("Utility class cannot be instantiated");
@@ -68,7 +59,7 @@ class RobotKeywordFileResolver {
     }
 
     static Collection<DefinedVariable> resolveVariables(PyFile pythonFile) {
-        if (isSystemLibrary(pythonFile)) {
+        if (RobotPyUtil.isSystemLibrary(pythonFile)) {
             return List.of();
         }
 
@@ -267,22 +258,6 @@ class RobotKeywordFileResolver {
             }
             keywords.add(new KeywordDto(method, libraryName, methodName, Arrays.asList(method.getParameterList().getParameters())));
         }
-    }
-
-    private static boolean isSystemLibrary(PsiFile psiFile) {
-        return CachedValuesManager.getCachedValue(psiFile, SYSTEM_PSI_FILE_KEY, () -> {
-            Module module = ModuleUtilCore.findModuleForPsiElement(psiFile);
-            if (module != null) {
-                Sdk sdk = PythonSdkUtil.findPythonSdk(module);
-                if (sdk != null) {
-                    VirtualFile[] roots = sdk.getRootProvider().getFiles(OrderRootType.CLASSES);
-                    VirtualFile fileVirtual = psiFile.getVirtualFile();
-                    boolean result = Arrays.stream(roots).anyMatch(root -> VfsUtilCore.isAncestor(root, fileVirtual, false));
-                    return Result.createSingleDependency(result, psiFile);
-                }
-            }
-            return Result.createSingleDependency(false, PsiModificationTracker.MODIFICATION_COUNT);
-        });
     }
 
     private static boolean isLibraryDecorated(PyClass pyClass) {